requirements.txt
dash
pandas
gunicorn
flask
dash-auth==1.3.2
requests
plotly.express
geopandas
load_dotenv